Move Form module to a separate file
